ibus-akaza/requirements.txt
skkdictutils>=0.0.2
jaconv>=0.2.4
PyGObject>=3.38.0
PyYAML==5.3.1
